requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.35
Werkzeug==3.0.4
itsdangerous==2.2.0
click==8.1.7
Flask-CORS==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
Pillow==10.2.0
psycopg2-binary==2.9.9
Flask-JWT-Extended==4.5.0
twilio==8.11.0